Implemented ITestable interface

+++ Common/TestFramework/IStatistics.cs
@@ -2,7 +2,7 @@
 {
     public interface IStatistics
     {
-        float Memory { get; set; }
+        long Memory { get; set; }
         long Duration { get; set; }
         double FramePerSecond { get; set; }
         string Description { get; set; }
